Add files via upload
This is a zipped up file of all the components required to run the add in. Most of the stuff should be in the addin.esriaddin file, however there are some additional addins which are located in the Install folder which writes "Hello World" in the combo-box and informs the user of the progress of the model when it is running.